moved common files to marl_utils directory
