+++ sumo_analysis.py
@@ -29,7 +29,7 @@ from sumo_custom_reward import MaxSpeedRewardFunction
 from sumo_custom_reward_avg_speed_limit import AverageSpeedLimitReward
 
 # Config Parser
-from MARLConfigParser import MARLConfigParser
+from marl_utils.MARLConfigParser import MARLConfigParser
 from rl_core.actor_critic import Actor
 from calculate_speed_control import CalculateSpeedError
